Add action application to verify results
