+++ planning.py
@@ -389,6 +389,31 @@ def three_block_tower():
                                            domain='Block(b) & Block(x)')],
                            domain='Block(A) & Block(B) & Block(C)')
 
+def logisticsPlanCustom(initial_state=None, goal_state=None):
+    if initial_state == None: 
+        initial_state = 'In(C1, R1) & In(C2, D1) & In(C3, D2) & In(R1, D1) & Holding(R1)'
+    if goal_state == None:
+        raise ValueError("Goal must be defined")
+
+    planning_problem = \
+    PlanningProblem(initial = initial_state,
+                    goals = goal_state,
+                    actions=[Action('PickUp(r, c, d)',
+                                    precond='In(r, d) & In (c, d) & ~Holding(r)',
+                                    effect='Holding(r) & ~In(c, d) & In(c, r)', 
+                                    domain='Robot(r) & Place(d) & Container(c)'),
+                             Action('PutDown(r, c, d)', 
+                                    precond='In(r, d) & In(c, r) & Holding(r)',
+                                    effect='~Holding(r) & ~In(c, r) & In(c, d)',
+                                    domain='Robot(r) & Place(d) & Container(c)'),
+                            Action('Move(r, d_start, d_end)',
+                                   precond='In(r,d_start)',
+                                   effect='~In(r, d_start) & In(r, d_end)',
+                                   domain='Robot(r) & Place(d_start) & Place(d_end)')],
+                domain='Container(C1) & Container(C2) & Container(C3) & Place(D1) & Place(D2) & Place(D3) & Robot(R1)')
+    
+    return planning_problem
+
 
 def simple_blocks_world():
     """
