+++ submission1_test.py
@@ -4,8 +4,6 @@ from planning import *
 from logic import *
 
 import pytest
-import submission1 as sub
-
 
 def test_blocksworld_manual():
     sbw = simple_blocks_world()
@@ -17,42 +15,41 @@ def test_blocksworld_manual():
     assert sbw.goal_test() == True
 
 
+def verify_solution(P):
+    sol = Linearize(P).execute()
+    print(sol)
+    assert P.goal_test() == False
+    for act in sol:
+        P.act(expr(act))
+    assert P.goal_test() == True
+
 def ptest_air_cargo():
     P = air_cargo()
-    assert isinstance(Linearize(P).execute(), list)
+    verify_solution(P)
 
 def test_spare_tire():
     P = spare_tire()
-    assert isinstance(Linearize(P).execute(), list)
-
-def test_three_block_tower():
+    verify_solution(P)
+    
+def ptest_three_block_tower():
     P = three_block_tower()
-    assert isinstance(Linearize(P).execute(), list)
+    verify_solution(P)
 
 def test_simple_blocks_world():
     P = simple_blocks_world()
-    assert isinstance(Linearize(P).execute(), list)
+    verify_solution(P)
 
 def test_shopping_problem():
     P = shopping_problem()
-    #print(GraphPlan(P).execute())
-    """
-    [[[PItem(Milk), PSells(SM, Milk), PSells(SM, Banana), PStore(HW), 
-    PItem(Banana), PStore(SM), Go(Home, HW), PSells(HW, Drill), 
-    PItem(Drill), Go(Home, SM)], 
-    [Buy(Drill, HW), Buy(Banana, SM), Buy(Milk, SM)]]]
-    """
-    assert isinstance(Linearize(P).execute(), list)
+    verify_solution(P)
 
 def test_socks_and_shoes():
     P = socks_and_shoes()
-    assert isinstance(Linearize(P).execute(), list)
+    verify_solution(P)
 
 def ptest_double_tennis_problem():
     P = double_tennis_problem()
-    #print(GraphPlan(P).execute())
-    assert isinstance(Linearize(P).execute(), list)
-
+    verify_solution(P)
 
 @pytest.mark.parametrize("goal_state", [
     "In(C1, D1)",
@@ -73,9 +70,8 @@ def ptest_double_tennis_problem():
 def test_logistics_plan_valid(goal_state):
     """These should yield a valid (non-crashing) plan, even if empty."""
     init = "In(C1, R1) & In(C2, D1) & In(C3, D2) & In(R1, D1) & Holding(R1)"
-    result = sub.logisticsPlanCustom(init, goal_state)
-    assert result is None or isinstance(result, list)
-
+    P = logisticsPlanCustom(init, goal_state)
+    verify_solution(P)
 
 @pytest.mark.parametrize("goal_state", [
     "In(C2, D3) & In(C3, D3)",
@@ -87,9 +83,10 @@ def test_logistics_plan_valid(goal_state):
 def ptest_logistics_plan_no_plan(goal_state):
     """These are known to have no valid plan."""
     init = "In(C1, R1) & In(C2, D1) & In(C3, D2) & In(R1, D1) & Holding(R1)"
-    result = sub.logisticsPlanCustom(init, goal_state)
+    P = logisticsPlanCustom(init, goal_state)
+    sol = Linearize(P).execute()
     # Depending on your GraphPlan, no-plan might return [] or None
-    assert result in ([], None)
+    assert sol in ([], None)
 
 
 @pytest.mark.parametrize("goal_state", [
@@ -99,8 +96,9 @@ def ptest_logistics_plan_kaboom(goal_state):
     """This case is known to cause planner explosion. Catch and mark as expected failure."""
     init = "In(C1, R1) & In(C2, D1) & In(C3, D2) & In(R1, D1) & Holding(R1)"
     try:
-        result = sub.logisticsPlanCustom(init, goal_state)
+        P = logisticsPlanCustom(init, goal_state)
+        sol = Linearize(P).execute()
         # If it actually returns, that's fine too
-        assert result is None or isinstance(result, list)
+        assert sol is None or isinstance(sol, list)
     except Exception:
         pytest.xfail("Known kaboom case – planner explosion or unsolvable problem")
